% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_spaces}
\alias{count_spaces}
\title{Count number of spaces preceding a character input}
\usage{
count_spaces(x, combine_single_spaces = FALSE)
}
\arguments{
\item{x}{A \code{character} input to be split.}

\item{combine_single_spaces}{A \code{logical} input. Should strings separated by only 1 space be recombined? Defaults to \code{FALSE}.}
}
\value{
A \code{data.frame}
}
\description{
For a given \code{character} input, split based on spaces (\code{" "}), and output a \code{dataframe}
with the number of preceding spaces before each word.
}
\author{
Matt Lewis
}
\keyword{internal}
