% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_space_width}
\alias{estimate_space_width}
\title{Estimate space width}
\usage{
estimate_space_width(x)
}
\arguments{
\item{x}{A \code{data.frame} or \code{tibble} passed by \code{split_pdf_text()}.}
}
\value{
A \code{data.frame} with a column for average space width added.
}
\description{
Estimate width of a space in a given font.
}
\author{
Matt Lewis
}
\keyword{internal}
