% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_pdf}
\alias{read_pdf}
\title{Read \code{.pdf} files into R}
\usage{
read_pdf(x)
}
\arguments{
\item{x}{\code{character}. File path to the \code{.pdf} file.}
}
\value{
A \code{list} object.
}
\description{
Reads \code{.pdf} files into R as text and does basic formatting including splitting paragraphs
}
\author{
Matt Lewis
}
