% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_pdf_text}
\alias{split_pdf_text}
\title{Split pdf data into text sections}
\usage{
split_pdf_text(x, size)
}
\arguments{
\item{x}{A \code{list} of \code{data.frames} as read by \code{pdftools::pdf_data(font_info = T)}.}

\item{size}{A \code{data.frame} as read by \code{pdftools::pdf_pagesize()}.}
}
\value{
A \code{list}.
}
\description{
For data read in by \code{pdftools::pdf_data()}, split into text sections based on x and y
positioning, height and width.
}
\author{
Matt Lewis
}
\keyword{internal}
